requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
django-crispy-forms==1.12.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
Markdown==3.3.4
Pillow==8.3.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
